+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to
 
 ### Added
 
+- Support Backup Codes for Multi-Factor Authentication
+  [937](https://github.com/OpenFn/Lightning/issues/937)
+- Log a warning in the console when the Editor/docs component is given latest
+  [#958](https://github.com/OpenFn/Lightning/issues/958)
+
 ### Changed
 
 ### Fixed